+++ NEWS.md
@@ -1,7 +1,11 @@
 # surveydown (development version)
 
+# surveydown 0.3.5
+
 - The database updating is simplified to only write to the database on each page turn and after the survey closes.
 - Bug fix: if you added new questions or values to the survey after creating the initial database table, those new values would not have been added to the table. Now they are added.
+- Now the server will only update the data upon page flip and survey closure in order to decrease data traffic.
+- `auto_scroll` changed to `FALSE` by default.
 
 # surveydown 0.3.4